Use NewCmd* pattern for search commands. Move util functions to util package.
